move mapper to mapper package
