+++ server/controller/src/main/java/ai/starwhale/mlops/domain/project/mapper/ProjectMapper.java
@@ -5,8 +5,9 @@
  * in accordance with the terms of the license agreement you entered into with StarWhale.ai.
  */
 
-package ai.starwhale.mlops.domain.project;
+package ai.starwhale.mlops.domain.project.mapper;
 
+import ai.starwhale.mlops.domain.project.ProjectEntity;
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
