+++ server/controller/src/main/java/ai/starwhale/mlops/domain/system/SystemService.java
@@ -9,6 +9,7 @@ package ai.starwhale.mlops.domain.system;
 
 import ai.starwhale.mlops.api.protocol.agent.AgentVO;
 import ai.starwhale.mlops.common.PageParams;
+import ai.starwhale.mlops.domain.system.mapper.AgentMapper;
 import com.github.pagehelper.PageHelper;
 import java.util.List;
 import java.util.stream.Collectors;
