+++ server/controller/src/main/java/ai/starwhale/mlops/reporting/ReportProcessorImp.java
@@ -13,8 +13,8 @@ import ai.starwhale.mlops.api.protocol.report.resp.ReportResponse;
 import ai.starwhale.mlops.domain.node.Node;
 import ai.starwhale.mlops.domain.system.Agent;
 import ai.starwhale.mlops.domain.system.AgentEntity;
-import ai.starwhale.mlops.domain.system.AgentMapper;
-import ai.starwhale.mlops.domain.task.TaskMapper;
+import ai.starwhale.mlops.domain.system.mapper.AgentMapper;
+import ai.starwhale.mlops.domain.task.mapper.TaskMapper;
 import ai.starwhale.mlops.domain.task.TaskStatus;
 import ai.starwhale.mlops.domain.task.bo.StagingTaskStatus;
 import ai.starwhale.mlops.domain.task.bo.Task;
@@ -25,15 +25,12 @@ import ai.starwhale.mlops.domain.task.LivingTaskStatusMachine;
 import ai.starwhale.mlops.api.protocol.report.resp.TaskTrigger;
 import ai.starwhale.mlops.domain.task.bo.TaskStatusStage;
 import ai.starwhale.mlops.schedule.CommandingTasksChecker;
-import ai.starwhale.mlops.schedule.TaskScheduler;
+import ai.starwhale.mlops.schedule.SWTaskScheduler;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.time.LocalDateTime;
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -51,7 +48,7 @@ public class ReportProcessorImp implements ReportProcessor {
 
     final LivingTaskStatusMachine livingTaskStatusMachine;
 
-    final TaskScheduler taskScheduler;
+    final SWTaskScheduler SWTaskScheduler;
 
     final TaskBoConverter taskBoConverter;
 
@@ -62,12 +59,12 @@ public class ReportProcessorImp implements ReportProcessor {
     final TaskMapper taskMapper;
 
     public ReportProcessorImp(CommandingTasksChecker commandingTasksChecker,
-        LivingTaskStatusMachine livingTaskStatusMachine, TaskScheduler taskScheduler,
+        LivingTaskStatusMachine livingTaskStatusMachine, SWTaskScheduler SWTaskScheduler,
         TaskBoConverter taskBoConverter, AgentMapper agentMapper, ObjectMapper jsonMapper,
         TaskMapper taskMapper) {
         this.commandingTasksChecker = commandingTasksChecker;
         this.livingTaskStatusMachine = livingTaskStatusMachine;
-        this.taskScheduler = taskScheduler;
+        this.SWTaskScheduler = SWTaskScheduler;
         this.taskBoConverter = taskBoConverter;
         this.agentMapper = agentMapper;
         this.jsonMapper = jsonMapper;
@@ -101,7 +98,7 @@ public class ReportProcessorImp implements ReportProcessor {
             return rebuildReportResponse(unProperTasks);
         }
         taskStatusChange(tasks);
-        final List<Task> toAssignTasks = taskScheduler.schedule(nodeInfo);
+        final List<Task> toAssignTasks = SWTaskScheduler.schedule(nodeInfo);
         final Collection<Task> toCancelTasks = livingTaskStatusMachine
             .ofStatus(new StagingTaskStatus(TaskStatus.CANCEL));
         scheduledTaskStatusChange(toAssignTasks,agentEntity);
