+++ server/controller/src/main/resources/mapper/BaseImageMapper.xml
@@ -7,7 +7,7 @@
   -->
 
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-<mapper namespace="ai.starwhale.mlops.domain.job.BaseImageMapper">
+<mapper namespace="ai.starwhale.mlops.domain.job.mapper.BaseImageMapper">
     <select id="listBaseImages" resultMap="baseImageResultMap">
         select id as image_id,
                image_name,
