+++ server/controller/src/main/resources/mapper/JobSWDSVersionMapper.xml
@@ -7,8 +7,8 @@
   -->
 
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-<mapper namespace="ai.starwhale.mlops.domain.job.JobSWDSVersionMapper">
-    <select id="listSWDSVersionsByJobId" resultMap="ai.starwhale.mlops.domain.swds.SWDatasetVersionMapper.dsVersionResultMap">
+<mapper namespace="ai.starwhale.mlops.domain.job.mapper.JobSWDSVersionMapper">
+    <select id="listSWDSVersionsByJobId" resultMap="ai.starwhale.mlops.domain.swds.mapper.SWDatasetVersionMapper.dsVersionResultMap">
         select dsv.id as dsv_id,
                dataset_id,
                owner_id,
