+++ server/controller/src/main/resources/mapper/TaskMapper.xml
@@ -7,7 +7,7 @@
   -->
 
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-<mapper namespace="ai.starwhale.mlops.domain.task.TaskMapper">
+<mapper namespace="ai.starwhale.mlops.domain.task.mapper.TaskMapper">
     <sql id="select_task">
         select t.id as task_id,
                t.task_uuid,
@@ -97,6 +97,6 @@
         <result property="swdsBlocks" column="task_swds_blocks" />
         <result property="createdTime" column="task_created_time" />
         <result property="modifiedTime" column="task_modified_time" />
-        <association property="agent" resultMap="ai.starwhale.mlops.domain.system.AgentMapper.agentResultMap" />
+        <association property="agent" resultMap="ai.starwhale.mlops.domain.system.mapper.AgentMapper.agentResultMap" />
     </resultMap>
 </mapper>
